requirements/production.txt
Django==1.11
psycopg2
prettyconf
dj-database-url
djangorestframework
django-mptt
djangorestframework-recursive